ALMACEN. (elaboraciones) mas cambios para suggestBox de item

+++ aon-gwt-api/src/main/java/com/esferalia/aon/gwt/api/client/warehouse/JsElaboration.java
@@ -61,7 +61,7 @@ public class JsElaboration extends JavaScriptObject {
 												return this.comments;
 												}-*/;
 
-	public final native Byte getSource() /*-{
+	public final native Integer getSource() /*-{
 											return this.source;
 											}-*/;
